test_outlook/docker/requirements.txt
flask
llama_index
python-dotenv
